Custom treatment of `Typeable` in super-classes.
It would appear that GHC "short-cuts" the solver when it encounters
super-class constraints that look like classes.

This means that the custom solvers in TcInteract do not work!
This does not seem quite right, but until we fix it, we have
an explicit check to pass on `Typeable` to the constraint solver.