CmmSwitch: Move table offset to code generation phase
Previously, if a switch statement would not start with 0, the Stg → Cmm
phase would offset the scrutinee to make the table zero-based. In order
to have CmmSwitch a bit higher level, this step is moved to the Cmm →
Assembly phase.

This also means that in the llvm backend, more is more
logic left to the LLVM compiler (which hopefully knows best how to
compile a switch statement).